Refactor party: Instead of having 2 pledge animation methods we just have one that figures out all the start and end values. Some new linez. Created
BottomCropImageView that performs a matrix scale with the bottom of the scaled view cut off."

+++ app/src/main/java/com/kickstarter/ui/data/ScreenLocation.kt
@@ -2,4 +2,4 @@ package com.kickstarter.ui.data
 
 import java.io.Serializable
 
-data class ScreenLocation(val x: Float, val y: Float, val height: Int, val width: Int) : Serializable
+data class ScreenLocation(val x: Float, val y: Float, val height: Float, val width: Float) : Serializable
